requirements.txt
streamlit
requests
beautifulsoup4
trafilatura
lxml
tldextract
requests-cache
readability-lxml
urllib3
fastapi
uvicorn
markdown
pdfminer.six
pyyaml
